spring-boot-project/spring-boot-cli/src/main/java/org/springframework/boot/cli/command/OptionParsingCommand.java


package org.springframework.boot.cli.command;

import java.util.Collection;

import org.springframework.boot.cli.command.options.OptionHandler;
import org.springframework.boot.cli.command.options.OptionHelp;
import org.springframework.boot.cli.command.status.ExitStatus;

/**
 * Base class for a {@link Command} that parse options using an {@link OptionHandler}.
 *
 * @author Phillip Webb
 * @author Dave Syer
 * @see OptionHandler
 */
public abstract class OptionParsingCommand extends AbstractCommand {

	private final OptionHandler handler;

	protected OptionParsingCommand(String name, String description,
			OptionHandler handler) {
		super(name, description);
		this.handler = handler;
	}

	@Override
	public String getHelp() {
		return this.handler.getHelp();
	}

	@Override
	public Collection<OptionHelp> getOptionsHelp() {
		return this.handler.getOptionsHelp();
	}

	@Override
	public final ExitStatus run(String... args) throws Exception {
		return this.handler.run(args);
	}

	protected OptionHandler getHandler() {
		return this.handler;
	}

}
